Add files via upload
Create Quiz.  
Version 1 simple - Allows randomization of questions and answers, but forces user to type in exact answers.
Version 2 - Added ability of user to select answers based on an answer number (no typing it in).
Version 3 - Added ability for the quiz creator to put all of the questions in a spreadsheet and load it in, rather than creating the questions in code.